% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netQDApredict.R
\name{netQDApredict}
\alias{netQDApredict}
\title{netQDA Classification Step}
\usage{
netQDApredict(TrainModel, X.new)
}
\arguments{
\item{TrainModel}{the output from netQDAtrain function.}

\item{X.new}{new data for classification task.}
}
\value{
Fisher.matrix Fisher's matrix

predClass predicted group membership

p1 predicted probability of group 1

p2 predicted probability of group 2
}
\description{
netQDA Classification Step
}
