% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netQDAtrain.R
\name{netQDAtrain}
\alias{netQDAtrain}
\title{netQDA Train Step}
\usage{
netQDAtrain(
  X,
  y,
  COV.list = NULL,
  COR.list = NULL,
  COV.scale.list = NULL,
  pair = c(1, 2),
  tau = 20,
  alpha = 0.7,
  nu = 50,
  d = 2,
  nb = 10,
  delta = 100,
  p1 = NULL,
  p2 = NULL
)
}
\arguments{
\item{X}{is a n by p matrix}

\item{y}{is a vector of length n}

\item{COV.list}{is a list with each element corresponding to the covariance matrix of each group.}

\item{COR.list}{is a list with each element corresponding to the correlation matrix of each group.}

\item{COV.scale.list}{is a list with each element corresponding to the covariance matrix of standadized data in each group.}

\item{pair}{the index for group. For example, pair=c(1,2), denotes the comparison between group 1 and 2}

\item{tau}{is the hyper-parameter for the number of strong differential expressed genes.}

\item{alpha}{is the hyper-parameter for the threshold for correlation matrix. We suggest using threshold no smaller than 0.7}

\item{nu}{is the hyper-parameter for the number of differential expressed (DE) genes.}

\item{d}{is the hyper-parameter depth/layer when searching for connected component based on the strong differential expressed (DE) gene.}

\item{nb}{is the hyper-parameter of number of features for each layer when searching for connected component based on the strong differential expressed gene.}

\item{delta}{is the hyper-parameter for the number of differential connected (DC)  genes.}

\item{p1}{is the probability of a data point belonging to group 1.}

\item{p2}{is the probability of a data point belonging to group 2. p1 + p2 = 1.}
}
\value{
try.meanDiff.firstTau the mean differences of data in group 1 and group 2

try.local.DCnetwork_list connected component of DC features based on DE features

try.local.DEnetwork_list connected component of DE features based on DE features

screenset.DC selected DC features

screenset.DE selected DE feautres

screenset.DE.iff if and only if condition values for DE features

screenset selected features (both DC and DE)

Omega.cis.mat.list Precision matrix of each connected component of selected features.

det.cov.mat determinants of covariance matrix of each group.

log.det.cov.mat log of determinants of covariance matrix of each group.

p.vec the probability of belong to group 1 and group 2.

pair the group index of the two groups.
}
\description{
netQDA Train Step
}
